Change patterns definition using doublestar package
