+++ internal/upload/types.go
@@ -1,5 +1,15 @@
 package upload
 
+// UploadFolderJob represents a job to upload all photos from the specified folder
+type UploadFolderJob struct {
+	FileTracker FileTracker
+
+	SourceFolder       string
+	CreateAlbum        bool
+	CreateAlbumBasedOn string
+	Filter             FileFilterer
+}
+
 // FileTracker represents a service to track already uploaded files.
 type FileTracker interface {
 	CacheAsAlreadyUploaded(filePath string) error
@@ -7,12 +17,8 @@ type FileTracker interface {
 	RemoveAsAlreadyUploaded(filePath string) error
 }
 
-// UploadFolderJob represents a job to upload all photos from the specified folder
-type UploadFolderJob struct {
-	FileTracker   FileTracker
-
-	SourceFolder       string
-	CreateAlbum        bool
-	CreateAlbumBasedOn string
-	Filter             *Filter
+// FileFilterer represents a way to implement include/exclude files filtering.
+type FileFilterer interface {
+	IsAllowed(path string) bool
+	IsExcluded(path string) bool
 }
